refactor: move JoinGraph into Optimizer/Utils
The join cost estimator is a helper component for the optimize-join-order
rule and belongs in Optimizer/Utils, but its interface is expressed in terms
of JoinGraph::Node and Edge. Moving JoinGraph out of Rule/ keeps the
dependency direction pointing from rules to utils.
