+++ arangod/Aql/Optimizer/Rule/OptimizeJoinOrder.h
@@ -21,82 +21,15 @@
 
 #pragma once
 
-#include <map>
+#include "Aql/Optimizer/Utils/JoinGraph.h"
+
 #include <memory>
-#include <string_view>
-#include <vector>
 
 namespace arangodb::aql {
-class EnumerateCollectionNode;
 class ExecutionNode;
 class ExecutionPlan;
 class Optimizer;
-struct AstNode;
 struct OptimizerRule;
-struct Variable;
-
-/// @brief a path of nested attribute accesses, e.g. `doc.a.b` -> {"a", "b"}.
-using AttributePath = std::vector<std::string_view>;
-
-/// @brief The join graph describes a maximal run of adjacent collection
-/// enumerations (`FOR`-loops) that are connected via equijoin conditions.
-struct JoinGraph {
-  struct Node {
-    EnumerateCollectionNode* executionNode;
-
-    /// @brief attribute paths that are constrained by a constant restriction
-    /// on this node (e.g. the `x` of `doc.x == 'foo'`).
-    std::vector<AttributePath> conditions;
-
-    explicit Node(EnumerateCollectionNode* executionNode)
-        : executionNode(executionNode) {}
-  };
-
-  struct Edge {
-    Node* from;
-    Node* to;
-
-    std::vector<AttributePath> fromAttributes;
-    std::vector<AttributePath> toAttributes;
-
-    Edge(Node* from, Node* to) : from(from), to(to) {}
-  };
-
-  /// @brief the vertex for a given out variable, or nullptr if the variable is
-  /// not part of this join graph.
-  auto nodeForVariable(Variable const* variable) -> Node*;
-
-  void addNode(EnumerateCollectionNode* en);
-
-  /// @brief record an equijoin condition `v.<vAttributes> == w.<wAttributes>`
-  /// as an edge between the vertices for `v` and `w`.
-  auto addJoinCondition(Variable const* v, AttributePath vAttributes,
-                        Variable const* w, AttributePath wAttributes) -> Edge&;
-
-  /// @brief record a predicate that could not be classified as an equijoin or a
-  /// constant restriction. It is left untouched in the plan and only kept here
-  /// for later optimizer stages.
-  void addResidual(AstNode const* node);
-
-  auto getEdgesForNode(Node* node) -> std::vector<Edge*>;
-
-  /// @brief partition the vertices into connected components (the graph may be
-  /// disconnected). Each returned vector holds the out variables of one
-  /// component. Order within and across components is unspecified.
-  [[nodiscard]] auto connectedComponents() const
-      -> std::vector<std::vector<Variable const*>>;
-
-  /// @brief true if this graph describes an actual join, i.e. it has at least
-  /// one equijoin edge.
-  [[nodiscard]] auto hasJoin() const noexcept -> bool { return !edges.empty(); }
-
-  std::map<Variable const*, Node> nodes;
-  std::vector<Edge> edges;
-  std::vector<AstNode const*> residuals;
-
- private:
-  auto ensureEdge(Variable const* v, Variable const* w) -> Edge&;
-};
 
 /// @brief build the join graph for the maximal run of adjacent collection
 /// enumerations that starts at `firstEnumeration`. Walks upwards (towards the
@@ -107,8 +40,7 @@ struct JoinGraph {
 auto buildJoinGraph(ExecutionPlan const* plan, ExecutionNode* firstEnumeration,
                     ExecutionNode*& next) -> JoinGraph;
 
-/// @brief the `optimize-join-order` optimizer rule. Only constructs the join
-/// graph(s) for now. Does not modify the plan.
+/// @brief the `optimize-join-order` optimizer rule.
 void optimizeJoinOrder(Optimizer*, std::unique_ptr<ExecutionPlan>,
                        OptimizerRule const&);
