+++ arangod/Aql/Optimizer/Utils/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(arango_aql PRIVATE
   AccessesCollectionVariable.cpp
   EnumeratePathsFilterMatcher.cpp
+  JoinGraph.cpp
   LateMaterializedCommon.cpp
   OptimizeTraversalPathVariable.cpp
 )
