+++ tests/Aql/OptimizeJoinOrderTest.cpp
@@ -19,21 +19,7 @@
 /// Copyright holder is ArangoDB GmbH, Cologne, Germany
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "gtest/gtest.h"
-
-#include "Aql/ExecutionNode/ExecutionNode.h"
-#include "Aql/ExecutionPlan.h"
-#include "Aql/Optimizer/Rule/OptimizeJoinOrder.h"
-#include "Aql/Query.h"
-#include "Aql/QueryString.h"
-#include "Aql/Variable.h"
-#include "Transaction/StandaloneContext.h"
-
-#include "velocypack/Parser.h"
-
-#include "../IResearch/common.h"
-#include "../Mocks/Servers.h"
-#include "Async/async.h"
+#include "JoinGraphTestHelper.h"
 
 #include <algorithm>
 #include <memory>
@@ -58,65 +44,8 @@ class OptimizeJoinOrderTest : public testing::Test {
     }
   }
 
-  // Prepares a plan for `query` without any optional optimizer rules, so that
-  // the plain EnumerateCollection / Filter nodes survive (this mirrors what the
-  // optimize-join-order rule sees at its position in the pipeline).
-  std::shared_ptr<Query> prepare(std::string const& query) const {
-    auto ctx = std::make_shared<transaction::StandaloneContext>(
-        server.getSystemDatabase(), transaction::OperationOriginTestCase{});
-    auto options =
-        velocypack::Parser::fromJson(R"({"optimizer":{"rules":["-all"]}})");
-    auto q = Query::create(std::move(ctx), QueryString(query), nullptr,
-                           QueryOptions(options->slice()));
-    waitForAsync(q->prepareQuery());
-    return q;
-  }
-
-  // Walks from the singleton towards the root and returns the first
-  // EnumerateCollection node (i.e. the bottom-most enumeration).
-  static ExecutionNode* firstEnumeration(const ExecutionPlan* plan) {
-    for (ExecutionNode* n = plan->root()->getSingleton(); n != nullptr;
-         n = n->getFirstParent()) {
-      if (n->getType() == ExecutionNode::ENUMERATE_COLLECTION) {
-        return n;
-      }
-    }
-    return nullptr;
-  }
-
-  static JoinGraph build(const Query& q) {
-    auto* plan = q.plan();
-    auto* first = firstEnumeration(plan);
-    EXPECT_NE(first, nullptr);
-    ExecutionNode* next = nullptr;
-    return buildJoinGraph(plan, first, next);
-  }
-
-  // Builds one JoinGraph per maximal run of adjacent enumerations, mirroring
-  // the spine walk of the optimizeJoinOrder rule itself.
-  static std::vector<JoinGraph> buildAll(const Query& q) {
-    auto* plan = q.plan();
-    std::vector<JoinGraph> graphs;
-    for (ExecutionNode* n = plan->root()->getSingleton(); n != nullptr;) {
-      if (n->getType() == ExecutionNode::ENUMERATE_COLLECTION) {
-        ExecutionNode* next = nullptr;
-        graphs.emplace_back(buildJoinGraph(plan, n, next));
-        n = next;
-      } else {
-        n = n->getFirstParent();
-      }
-    }
-    return graphs;
-  }
-
-  static JoinGraph::Node const* nodeByName(JoinGraph const& g,
-                                           std::string_view name) {
-    for (auto const& [var, node] : g.nodes) {
-      if (var->name == name) {
-        return &node;
-      }
-    }
-    return nullptr;
+  std::shared_ptr<Query> prepare(std::string const& query) {
+    return prepareJoinPlan(server, query);
   }
 };
 }  // namespace
@@ -125,7 +54,7 @@ TEST_F(OptimizeJoinOrderTest, linear_three_way_chain) {
   auto q = prepare(
       "FOR a IN c1 FOR b IN c2 FILTER a.x == b.y "
       "FOR c IN c3 FILTER b.z == c.w RETURN [a, b, c]");
-  auto g = build(*q);
+  auto g = buildGraph(*q);
 
   EXPECT_EQ(g.nodes.size(), 3u);
   EXPECT_EQ(g.edges.size(), 2u);
@@ -136,7 +65,7 @@ TEST_F(OptimizeJoinOrderTest, linear_three_way_chain) {
 
 TEST_F(OptimizeJoinOrderTest, equijoin_edge_records_attribute_paths) {
   auto q = prepare("FOR a IN c1 FOR b IN c2 FILTER a.x == b.y RETURN [a, b]");
-  auto g = build(*q);
+  auto g = buildGraph(*q);
 
   ASSERT_EQ(g.edges.size(), 1u);
   auto const& e = g.edges.front();
@@ -156,7 +85,7 @@ TEST_F(OptimizeJoinOrderTest, constant_restriction_becomes_node_condition) {
   auto q = prepare(
       "FOR a IN c1 FOR b IN c2 FILTER a.x == b.y "
       "FILTER a.k == 'v' RETURN [a, b]");
-  auto g = build(*q);
+  auto g = buildGraph(*q);
 
   EXPECT_EQ(g.nodes.size(), 2u);
   EXPECT_EQ(g.edges.size(), 1u);
@@ -173,7 +102,7 @@ TEST_F(OptimizeJoinOrderTest, non_equijoin_predicate_becomes_residual) {
   auto q = prepare(
       "FOR a IN c1 FOR b IN c2 FILTER a.x == b.y "
       "FILTER a.p < b.q RETURN [a, b]");
-  auto g = build(*q);
+  auto g = buildGraph(*q);
 
   EXPECT_EQ(g.edges.size(), 1u);
   EXPECT_FALSE(g.residuals.empty());
@@ -182,7 +111,7 @@ TEST_F(OptimizeJoinOrderTest, non_equijoin_predicate_becomes_residual) {
 TEST_F(OptimizeJoinOrderTest, id_is_remapped_to_key) {
   auto q =
       prepare("FOR a IN c1 FOR b IN c2 FILTER a._id == b._id RETURN [a, b]");
-  auto g = build(*q);
+  auto g = buildGraph(*q);
 
   ASSERT_EQ(g.edges.size(), 1u);
   auto const& e = g.edges.front();
@@ -197,7 +126,7 @@ TEST_F(OptimizeJoinOrderTest, disconnected_graph_has_two_components) {
   auto q = prepare(
       "FOR a IN c1 FOR b IN c2 FILTER a.x == b.y "
       "FOR c IN c3 FOR d IN c1 FILTER c.x == d.y RETURN [a, b, c, d]");
-  auto g = build(*q);
+  auto g = buildGraph(*q);
 
   EXPECT_EQ(g.nodes.size(), 4u);
   EXPECT_EQ(g.edges.size(), 2u);
@@ -206,7 +135,7 @@ TEST_F(OptimizeJoinOrderTest, disconnected_graph_has_two_components) {
 
 TEST_F(OptimizeJoinOrderTest, single_enumeration_has_no_join) {
   auto q = prepare("FOR a IN c1 FILTER a.x == 1 RETURN a");
-  auto g = build(*q);
+  auto g = buildGraph(*q);
 
   EXPECT_EQ(g.nodes.size(), 1u);
   EXPECT_TRUE(g.edges.empty());
@@ -221,7 +150,7 @@ TEST_F(OptimizeJoinOrderTest, separate_runs_produce_separate_graphs) {
       "FOR a IN c1 FOR b IN c2 FILTER a.x == b.y "
       "SORT a.x "
       "FOR c IN c3 FOR d IN c1 FILTER c.x == d.y RETURN [a, b, c, d]");
-  auto graphs = buildAll(*q);
+  auto graphs = buildAllGraphs(*q);
 
   ASSERT_EQ(graphs.size(), 2u);
   for (auto const& g : graphs) {
